requirements.txt
autopep8==1.5.4
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pycodestyle==2.6.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
toml==0.10.2
xlrd==1.2.0
